@@ -28,12 +28,12 @@
 	"minimum-stability": "dev",
 	"prefer-stable": true,
 	"require": {
-		"php": ">=7.2",
-		"icanboogie/activerecord": "^5.0"
+		"php": ">=8.0",
+		"icanboogie/activerecord": "^6.0"
 	},
 	"require-dev": {
 		"icanboogie/event": "^4.0",
-		"phpunit/phpunit": "^8.5"
+		"phpunit/phpunit": "^9.5"
 	},
 	"autoload": {
 		"psr-4": {
